requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Django==3.2.8
djangorestframework==3.12.4
gingerit==0.8.2
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
nltk==3.6.5
pytz==2021.3
regex==2021.10.23
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
whitenoise==5.3.0
